replace simd with R32F32SFloat and optimize

+++ src/TTauri/GUI/PipelineBox.cpp
@@ -30,8 +30,8 @@ void PipelineBox::drawInCommandBuffer(vk::CommandBuffer commandBuffer)
 
     commandBuffer.bindVertexBuffers(0, tmpVertexBuffers, tmpOffsets);
 
-    pushConstants.windowExtent = { extent.width , extent.height };
-    pushConstants.viewportScale = { 2.0 / extent.width, 2.0 / extent.height };
+    pushConstants.windowExtent = { numeric_cast<int>(extent.width) , numeric_cast<int>(extent.height) };
+    pushConstants.viewportScale = { 2.0f / extent.width, 2.0f / extent.height };
     commandBuffer.pushConstants(
         pipelineLayout,
         vk::ShaderStageFlagBits::eVertex | vk::ShaderStageFlagBits::eFragment,
