+++ src/TTauri/GUI/PipelineImage.cpp
@@ -33,10 +33,10 @@ void PipelineImage::drawInCommandBuffer(vk::CommandBuffer commandBuffer)
 
     commandBuffer.bindVertexBuffers(0, tmpVertexBuffers, tmpOffsets);
 
-    pushConstants.windowExtent = { extent.width , extent.height };
-    pushConstants.viewportScale = { 2.0 / extent.width, 2.0 / extent.height };
+    pushConstants.windowExtent = { numeric_cast<int>(extent.width) , numeric_cast<int>(extent.height) };
+    pushConstants.viewportScale = { 2.0f / extent.width, 2.0f / extent.height };
     pushConstants.atlasExtent = { DeviceShared::atlasImageWidth, DeviceShared::atlasImageHeight };
-    pushConstants.atlasScale = { 1.0 / DeviceShared::atlasImageWidth, 1.0 / DeviceShared::atlasImageHeight };
+    pushConstants.atlasScale = { 1.0f / DeviceShared::atlasImageWidth, 1.0f / DeviceShared::atlasImageHeight };
     commandBuffer.pushConstants(
         pipelineLayout,
         vk::ShaderStageFlagBits::eVertex | vk::ShaderStageFlagBits::eFragment,
