+++ src/TTauri/GUI/PipelineSDF.cpp
@@ -31,8 +31,8 @@ void PipelineSDF::drawInCommandBuffer(vk::CommandBuffer commandBuffer)
 
     commandBuffer.bindVertexBuffers(0, tmpVertexBuffers, tmpOffsets);
 
-    pushConstants.windowExtent = { extent.width , extent.height };
-    pushConstants.viewportScale = { 2.0 / extent.width, 2.0 / extent.height };
+    pushConstants.windowExtent = { numeric_cast<int>(extent.width) , numeric_cast<int>(extent.height) };
+    pushConstants.viewportScale = { 2.0f / extent.width, 2.0f / extent.height };
     pushConstants.subpixelOrientation = static_cast<int>(window.subpixelOrientation);
 
     commandBuffer.pushConstants(
